requirements.txt
pyonmttok==1.*
tensorflow-gpu==1.11.0
